% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{gGrowthCurve_example}
\alias{gGrowthCurve_example}
\alias{gGrowthCurve_example_dir}
\title{gGrowthCurve_example}
\usage{
gGrowthCurve_example(path = NULL)

gGrowthCurve_example_dir()
}
\arguments{
\item{path}{Name of the file. If \code{NULL}, it will list all files in \code{extdata}.}
}
\description{
gGrowthCurve includes example data in its \code{inst/extdata} directory.
}
\examples{
# list all files
gGrowthCurve_example()
}
