% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_import.R
\name{import_metadata}
\alias{import_metadata}
\title{import_metadata}
\usage{
import_metadata(file, well_id = "well_id", multi_plate = FALSE,
  plate_id = NULL, ...)
}
\arguments{
\item{file}{Metadata file to be imported. Must contain a column containing well IDs (A1, A2...)}

\item{well_id}{The name of the column in your metadata table with locations on the plate}

\item{multi_plate}{If you will be using multiple plates, set this to TRUE}

\item{plate_id}{If you are analyzing multiple plates, this variable is the name of the column in your metadata table where you store plate_ids}

\item{...}{Additional parameters for for data.table::fread()}
}
\description{
This function allows users to import metadata that will be used for downstream analyses.
}
